Create MainActivity.java
This is the module where the bar code is scanned and the result is displayed on a Toast. I have the search text box for searching by name as well. The next step is the search parameter is passed to the next view. Will be handled at the integration stage. Initial test on a device was successful. Needs more testing after integration.